Added new function to get a list of IAs who have logged in since a specific date. Refactor location of internal reporting functions. Fix function which gets candidates who have logged in since a specific date.
