+++ nodeScripts/temp/testGetIndependentAssessorsEmails.js
@@ -1,16 +1,18 @@
 'use strict';
 
 /**
- * Get a list of candidates who have signed in after a specific date
+ * Get a list of candidates who have signed in after a specific date and export to csv
  */
 
 import { app, db, auth } from '../shared/admin.js';
-import initUserRoles from '../../functions/actions/userRoles.js';
+import initInternalReporting from '../../functions/actions/internalReporting.js';
+
+
 import path  from 'path';
 
 import ExcelJS from 'exceljs';
 
-const { getIndependentAssessorsSinceSignInDate } = initUserRoles(db, auth);
+const { getIndependentAssessorsSinceSignInDate } = initInternalReporting(db, auth);
 
 const exportToExcel = async (data, filePath) => {
   const workbook = new ExcelJS.Workbook();
